fix(ci): stabilize quality gates
