+++ packages/app/src/components/patterns/field-block.tsx
@@ -1,6 +1,6 @@
 import { navetSpacingTokens, navetTypographyTokens } from '@navet/app/components/system/tokens';
 import { cn } from '@navet/app/components/ui/utils';
-import { useTheme } from '@navet/app/hooks';
+import { useTheme } from '@navet/app/hooks/use-theme';
 import type { ReactNode } from 'react';
 
 export interface FieldBlockProps {
