+++ packages/app/src/components/primitives/button.tsx
@@ -9,7 +9,6 @@ import {
 } from '@navet/app/components/system/tokens';
 import { cn } from '@navet/app/components/ui/utils';
 import { useTheme } from '@navet/app/hooks/use-theme';
-import { Loader2 } from 'lucide-react';
 import { type ButtonHTMLAttributes, forwardRef, type ReactNode } from 'react';
 
 export interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
@@ -126,7 +125,18 @@ export const Button = forwardRef<HTMLButtonElement, ButtonProps>(function Button
       }}
     >
       {loading ? (
-        <Loader2 className={`${navetIconSizeTokens.sm} animate-spin`} aria-hidden="true" />
+        <svg
+          viewBox="0 0 24 24"
+          fill="none"
+          stroke="currentColor"
+          strokeWidth="2"
+          strokeLinecap="round"
+          strokeLinejoin="round"
+          className={`${navetIconSizeTokens.sm} animate-spin`}
+          aria-hidden="true"
+        >
+          <path d="M21 12a9 9 0 1 1-6.2-8.56" />
+        </svg>
       ) : iconOnly ? (
         iconContent
       ) : (
